servers/mcp-neo4j-vector-graphrag/pyproject.toml
[project]
name = "mcp-neo4j-vector-graphrag"
version = "0.3.1"
description = "A simple Neo4j MCP server using graphrag VectorCypher retriever"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.10.5",
    "neo4j-graphrag>=1.9.1",
    "pydantic>=2.10.1",
    "langchain-openai>=0.3.30"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pyright>=1.1.389",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.3",
    "ruff>=0.11.5",
    "testcontainers[neo4j]>=4.10.0",
    "aiohttp>=3.8.0"
]

[project.scripts]
mcp-neo4j-vector-graphrag = "mcp_neo4j_vector_graphrag:main"
